feat: get week goal amount
