+++ src/main/java/graduation/spendiary/domain/goal/GoalMonthService.java
@@ -1,15 +1,15 @@
 package graduation.spendiary.domain.goal;
 
 import graduation.spendiary.domain.DatabaseSequence.SequenceGeneratorService;
-import graduation.spendiary.domain.spendingWidget.SpendingWidgetDto;
-import graduation.spendiary.domain.spendingWidget.SpendingWidgetRepository;
 import graduation.spendiary.domain.spendingWidget.SpendingWidgetService;
 import graduation.spendiary.exception.NoSuchContentException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.temporal.TemporalAdjusters;
+import java.time.temporal.WeekFields;
 import java.util.List;
 import java.util.Optional;
 import java.util.stream.Collectors;
@@ -17,7 +17,9 @@ import java.util.stream.Collectors;
 @Service
 public class GoalMonthService {
     @Autowired
-    private GoalMonthRepository repo;
+    private GoalMonthRepository monthRepo;
+    @Autowired
+    private GoalWeekRepository weekRepo;
     @Autowired
     private SpendingWidgetService spendingWidgetService;
     @Autowired
@@ -40,14 +42,14 @@ public class GoalMonthService {
     }
 
     public List<GoalDto> getAll(String userId) {
-        return repo.findByUser(userId).stream()
+        return monthRepo.findByUser(userId).stream()
                 .map(this::getDto)
                 .collect(Collectors.toList());
     }
 
     public GoalDto getById(long id) {
-        Optional<GoalMonth> goalMonth = repo.findById(id);
-        if (!repo.existsById(id))
+        Optional<GoalMonth> goalMonth = monthRepo.findById(id);
+        if (!monthRepo.existsById(id))
             throw new NoSuchContentException();
         return this.getDto(goalMonth.get());
     }
@@ -59,7 +61,7 @@ public class GoalMonthService {
         LocalDate start = LocalDate.now().with(TemporalAdjusters.firstDayOfMonth());;
         LocalDate end = LocalDate.now().with(TemporalAdjusters.lastDayOfMonth());;
 
-        if(repo.findByUserAndDate(userId, start).isEmpty()){
+        if(monthRepo.findByUserAndDate(userId, start).isEmpty()){
 
             goalMonth.setId(SequenceGeneratorService.generateSequence(GoalMonth.SEQUENCE_NAME));
             goalMonth.setStart(start);
@@ -67,7 +69,7 @@ public class GoalMonthService {
             goalMonth.setUser(userId);
             goalMonth.setState(state);
             goalMonth.setMonth(start.getMonthValue());
-            repo.save(goalMonth);
+            monthRepo.save(goalMonth);
             return true;
         }else return false;
 
@@ -82,5 +84,38 @@ public class GoalMonthService {
          * total_cost >= goal_amount => achieved
          */
         return true;
+    }
+
+
+    public GoalWeek getGoalWeekById(Long id) {
+        Optional<GoalWeek> goalWeeks = weekRepo.findById(id);
+        if (goalWeeks.isEmpty()) throw new NoSuchContentException();
+        return goalWeeks.get();
+    }
+
+    public boolean weekGoal(Long monthId, GoalWeek goalWeek) {
+        // month 안에 있는 월간 목표 가져오기
+        GoalMonth goalMonth = monthRepo.findById(monthId).get();
+        long monthAmount = goalMonth.getAmount();
+        long weekAmountSum = goalMonth.getWeekIds().stream()
+                .map(this::getGoalWeekById)
+                .map(GoalWeek::getAmount)
+                .mapToLong(Long::longValue).sum();
+
+
+
+        int whatWeek = LocalDate.now().get(WeekFields.ISO.weekOfMonth());
+        LocalDate start = LocalDate.now().with(TemporalAdjusters.dayOfWeekInMonth(whatWeek - 1, DayOfWeek.MONDAY)); //특정 주차의 월요일 날짜
+        LocalDate end = LocalDate.now().with(TemporalAdjusters.dayOfWeekInMonth(whatWeek, DayOfWeek.SUNDAY)); //일요일 날짜
+        if (monthRepo.findByUserAndDate(monthId, start).isEmpty()) {
+            goalWeek.setId(SequenceGeneratorService.generateSequence(GoalMonth.SEQUENCE_NAME));
+            goalWeek.setStart(start);
+            goalWeek.setEnd(end);
+            goalWeek.setGoalMonth(monthId);
+            goalWeek.setState(state);
+            goalWeek.getAmount()
+            monthRepo.save(goalWeek);
+            return true;
+        } else return false;
     }
 }
